heroku Dev testing

+++ server/schemas/resolvers.js
@@ -1,4 +1,4 @@
-const { User, Capsule, Post } = require("../models");
+const { User, Capsule, Post, Payment } = require("../models");
 const { AuthenticationError } = require("apollo-server-express");
 const { createWriteStream } = require("fs");
 const path = require("path");
@@ -8,7 +8,6 @@ const { signToken, authMiddleware } = require("../utils/auth");
 const { ObjectId } = require("mongodb");
 require("dotenv").config();
 const bcrypt = require("bcryptjs");
-
 
 cloudinary.config({
   cloud_name: process.env.CLOUDINARY_NAME,
@@ -25,6 +24,16 @@ const resolvers = {
     getCapsulesDev: async () => {
       return await Capsule.find({});
     },
+    getCapsules: async (parent, args, context) => {
+      if (context.user) {
+        const user = await User.findOne({ _id: context.user._id }).populate(
+          "capsules"
+        );
+        return user.capsules;
+      }
+      throw new AuthenticationError("Authentication error");
+    },
+
     me: async (parent, args, context) => {
       if (context.user) {
         return await User.findOne({ _id: context.user._id });
@@ -51,7 +60,7 @@ const resolvers = {
         return null; // Post not found
       }
 
-      console.log("post found", post);
+      console.log("post found");
 
       return post;
     },
@@ -60,12 +69,14 @@ const resolvers = {
     //!! ADD ATTENDEES  and req.session.name saved (maybe token and not session)
 
     // Create a capsule with a title and date by a logged in user
-    createCapsule: async (parent, { title, date }, context) => {
+    createCapsule: async (parent, { title, date, location }, context) => {
       if (context.user) {
+        console.log("context.user", context.user);
         const capsule = await Capsule.create({
           title,
           date,
-          owner: context.user._id,
+          location,
+          owner: context.user.username,
         });
         await User.findOneAndUpdate(
           { _id: context.user._id },
@@ -79,36 +90,23 @@ const resolvers = {
 
     // Add a post to a capsule by a logged in user
     uploadPost: async (parent, { file }) => {
-        try {
-          console.log("uploading post...", file)
-          // Upload the image to Cloudinary
-          const uploadResult = await cloudinary.uploader.upload(file, {
-            folder: 'wedloc', // Specify the folder in Cloudinary
-          });
-  
-          console.log("uploadResult", uploadResult)
-          // Return the result of the Cloudinary upload
-          return {
-            public_id: uploadResult.public_id,
-            secure_url: uploadResult.secure_url,
-          };
-        } catch (error) {
-          throw new Error('Error uploading image: ' + error.message);
-        }
-      },
+      try {
+        console.log("uploading post...", file);
+        // Upload the image to Cloudinary
+        const uploadResult = await cloudinary.uploader.upload(file, {
+          folder: "wedloc", // Specify the folder in Cloudinary
+        });
 
-
-    //   await Capsule.findOneAndUpdate(
-    //     { _id: capsuleId },
-    //     { $addToSet: { posts: post._id } }
-    //   );
-    //   await User.findOneAndUpdate(
-    //     { _id: context.user._id },
-    //     { $addToSet: { posts: post._id } }
-    //   );
-    //   return post;
-    // },
-    // Delete a post by a logged in user
+        console.log("uploadResult", uploadResult);
+        // Return the result of the Cloudinary upload
+        return {
+          public_id: uploadResult.public_id,
+          secure_url: uploadResult.secure_url,
+        };
+      } catch (error) {
+        throw new Error("Error uploading image: " + error.message);
+      }
+    },
     deletePost: async (parent, { postId }, context) => {
       if (context.user) {
         const post = await Post.findOneAndDelete({
@@ -140,13 +138,17 @@ const resolvers = {
       return newChat;
     },
 
-    addUser: async (parent, { username, email, password, firstName, lastName }) => {
+    addUser: async (
+      parent,
+      { username, email, password, firstName, lastName }
+    ) => {
       try {
-        const user = await User.create({ username, 
-          email, 
-          password, 
-          firstName, 
-          lastName 
+        const user = await User.create({
+          username,
+          email,
+          password,
+          firstName,
+          lastName,
         });
         const token = signToken(user);
         return { token, user };
@@ -156,7 +158,11 @@ const resolvers = {
       }
     },
 
-    updateUser: async (parent, { firstName, lastName, email, profilePic }, context) => {
+    updateUser: async (
+      parent,
+      { firstName, lastName, email, profilePic },
+      context
+    ) => {
       const contextUserId = context.user._id;
       try {
         const updatedUser = await User.findOneAndUpdate(
@@ -167,8 +173,7 @@ const resolvers = {
         console.log("updatedUser", updatedUser);
         console.log("contextUserId", contextUserId);
         return { updatedUser };
-      }
-      catch (error) {
+      } catch (error) {
         console.error("Error updating user:", error);
         throw new Error("Error updating user");
       }
@@ -176,24 +181,20 @@ const resolvers = {
     deleteUser: async (parent, { username }, context) => {
       const contextUserId = context.user._id;
       try {
-        const deletedUser = await User.findOneAndDelete(
-          { _id: contextUserId }
-        );
+        const deletedUser = await User.findOneAndDelete({ _id: contextUserId });
         console.log("deletedUser", deletedUser);
         console.log("contextUserId", contextUserId);
         return { deletedUser };
-      }
-      catch (error) {
+      } catch (error) {
         console.error("Error deleting user:", error);
         throw new Error("Error deleting user");
       }
     },
-    
+
     devDelUser: async (parent, { userId }) => {
       const user = await User.findOneAndDelete({ _id: userId });
-      console.log('user deleted', user);
+      console.log("user deleted", user);
     },
-
 
     login: async (parent, { username, password }) => {
       console.log("hit login");
@@ -210,6 +211,26 @@ const resolvers = {
       console.log(token, user);
       return { token, user };
     },
+
+    addPayment: async (
+      _,
+      { userId, chargeId, amount, currency, description }
+    ) => {
+      try {
+        const payment = new Payment({
+          userId,
+          chargeId,
+          amount,
+          currency,
+          description,
+        });
+        await payment.save();
+        return payment;
+      } catch (err) {
+        throw new Error(err);
+      }
+    },
+
     uploadFile: async (_, { file }) => {
       try {
         const { createReadStream, filename, mimetype } = await file;
@@ -226,7 +247,6 @@ const resolvers = {
               else resolve(result);
             }
           );
-
           fileStream.pipe(cloudStream);
         });
 
@@ -240,6 +260,79 @@ const resolvers = {
       } catch (error) {
         throw new Error(`Failed to upload file: ${error}`);
       }
+    },
+
+    upVote: async (parent, { capsuleId, postId }) => {
+      const capsuleIdObject = new ObjectId(capsuleId);
+      const postIdObject = new ObjectId(postId);
+
+      const cap = await Capsule.findOneAndUpdate(
+        { _id: capsuleIdObject, "posts._id": postIdObject },
+        { $inc: { "posts.$.upVotes": 1 } },
+        { new: true }
+      );
+
+      if (!cap) {
+        console.log("post not found");
+        return null;
+      }
+
+      const upvotes = await cap.posts.find(
+        (post) => post._id.equals(postIdObject)
+      );
+      // cap now contains the updated document
+      console.log("post found: " + upvotes.upVotes);
+
+      return upvotes;
+    },
+
+    downVote: async (parent, { capsuleId, postId }) => {
+      const capsule = await Capsule.findById({ _id: capsuleId });
+
+      const postIdObject = new ObjectId(postId);
+      const post = capsule.posts.find((post) => post._id.equals(postIdObject));
+      if (!post) {
+        console.log("post not found");
+        return null;
+      }
+
+      console.log("post found");
+
+      const upVote = post.upVotes + 1;
+
+      const updatedPost = await Capsule.findOneAndUpdate(
+        { _id: capsuleId, "posts._id": postIdObject },
+        { $set: { "posts.$.downVotes": upVote } },
+        { new: true }
+      );
+
+      return updatedPost;
+    },
+
+    addComment: async (parent, { capsuleId, postId, text, author }) => {
+      const capsule = await Capsule.findById({ _id: capsuleId });
+
+      const postIdObject = new ObjectId(postId);
+      const post = capsule.posts.find((post) => post._id.equals(postIdObject));
+      if (!post) {
+        console.log("post not found");
+        return null;
+      }
+
+      console.log("post found");
+
+      const newComment = {
+        text,
+        author,
+      };
+
+      const updatedPost = await Capsule.findOneAndUpdate(
+        { _id: capsuleId, "posts._id": postIdObject },
+        { $push: { "posts.$.comments": newComment } },
+        { new: true }
+      );
+
+      return updatedPost;
     },
   },
 };
